fix(package): update deps (#37)
This includes type conversion code from Vis Util. It was removed there
because I thought it was used only by Vis Timeline and was causing
multiple copies of Moment to be present in bundles. Moving it here is
still an improvement (only one Moment in all libs except Timeline and
two in Timeline) but a better solution should be found. See
https://github.com/orgs/visjs/teams/maintainer/discussions/3 for more.